Add try catch to program. Add back lost try catch when creating account

+++ LifeCycleBank/Bank.cs
@@ -131,7 +131,7 @@ namespace LifeCycleBank
                 return "false";
         }
 
-        public string CreateAccount(ICustomer customerId, int balance)
+        public string CreateAccount(ICustomer customerId, decimal balance)
         {
             try
             {
